Move update_trust_score and update_packages_vulnerabilities to common.py. Move Redis connection settings to settings.base. Hold locks for shorter period of time when fetching CVEs for Ubuntu and Debian. ubuntu_cve: Use logging instead of print. ubuntu_cve: Move imports to the top. ubuntu_cve: Use Redis lock.
